requirements.txt
opencv-python==4.5.4.60
torch==1.10.1+cu102
numpy==1.22.3
fastapi==0.75.0
starlette==0.17.1
torchvision==0.11.2+cu102
PyYAML==5.4.1
gdown==4.4.0
pydantic==1.9.0
requests==2.27.1
streamlit==1.7.0
Pillow==9.0.1
sklearn==0.0
scikit-learn==0.24.2
wandb==0.12.14